MDL-88547 tiny_premium: fix self-hosted detection in get_api_key
get_config() always returns config values as strings, but usecloud
was computed with a strict (!==) comparison against the int
manager::PACKAGE_SELF_HOSTED constant. "2" !== 2 is always true in
PHP, so usecloud was always true regardless of the configured plugin
source, and self-hosted sites always fell back to the Tiny Cloud CDN.

Cast the config value to int before comparing. Adds a PHPUnit test
covering both the cloud and self-hosted plugin_source values.

Reported during peer review of this issue.

+++ public/lib/editor/tiny/plugins/premium/classes/external/get_api_key.php
@@ -62,7 +62,7 @@ class get_api_key extends external_api {
         self::validate_context($context);
         return [
             'apikey' => get_config('tiny_premium', 'apikey'),
-            'usecloud' => get_config('tiny_premium', 'plugin_source') !== manager::PACKAGE_SELF_HOSTED,
+            'usecloud' => (int) get_config('tiny_premium', 'plugin_source') !== manager::PACKAGE_SELF_HOSTED,
         ];
     }
